Add support for Player concept
Players (either computers or users) are responsible for choosing the actions each character takes.

To support this the Battle is now between players not parties.

The Party class is now a shell with no code, only containing a List<Character>

So far the computer player always tells the character to perform their 0th action, which is DoNothing.

This completes the "Core Game: Actions and Players" assigment